renamed areachart -> customchart
