+++ electron/ui/src/components/CustomChart/CustomChart.js
