+++ electron/ui/src/tests/customchart.test.js
@@ -1,5 +1,5 @@
 import { render, screen } from '@testing-library/react';
-import AreaChart from "../components/AreaChart/AreaChart"
+import CustomChart from "../components/CustomChart/CustomChart"
 import mockScenario from './data/mockScenario.json'
 import * as React from 'react'
 
@@ -8,7 +8,7 @@ const mockCategory1 = "CompletionsDemand";
 const mockCategory2 = "v_F_Piped_dict";
 
 test('test data input area chart', () => {
-    render( <AreaChart
+    render( <CustomChart
             input
             category={"CompletionsPads"}
             data={mockScenario.data_input.df_parameters[mockCategory1]} 
@@ -23,7 +23,7 @@ test('test data input area chart', () => {
 })
 
 test('test model results area chart', () => {
-    render( <AreaChart
+    render( <CustomChart
             data={mockScenario.results.data[mockCategory2]} 
             title={mockCategory2}
             xaxis={{titletext: "X axis title"}}
